requirements.txt
streamlit
tensorflow-cpu==2.13.0
opencv-python-headless
numpy==1.23.5
Pillow
gdown
keras==2.11
